zebra: add new dataplane operation
Add a new value DPLANE_OP_PROVIDER_REFRESH to enum dplane_op_e.
The intent of this value is to allow providers to request zebra
sending its routing state. This is useful to resynchronize zebra's
state with that of a remote dataplane in case the communication
between the two was interrrupted, or the remote dataplane, if
software, was restarted.

Signed-off-by: Fredi Raspall <fredi@githedgehog.com>