zebra: add method to request refresh
Add method to allow providers to issue a refresh request to the
dataplane subsystem and emit a log in the latter when one such
request is processed. Subsequent patches will implement the
handling of such a request. The method has a unint32_t argument
so that providers can indicate their id. This is, however, just
for informational purposes.

Signed-off-by: Fredi Raspall <fredi@githedgehog.com>