zebra: avoid kernel errors on refresh
As it stands, the dataplane refresh support aims at retransfering
most of the routing state to a remote dataplane and handling the
responses with the original implementation such that if the
refresh of a route failed in the dataplane (e.g. because of lack
of resources), that failure would be reflected in zebra's fib
state for that route. However, the dplane refresh support allows
refreshing interface and addressing state too. However, unlike
with routes, the processing of responses to refresh interfaces or
addresses causes the kernel provider to complain about extant
interfaces or addresses. E.g.

 ZEBRA: [WVJCK-PPMGD][EC 4043309093] netlink-dp (NS 0)
   error: File exists, type=RTM_NEWADDR(20)...

To avoid those errors, and given that the nature of kernel inter-
face and address data may be informational to a dataplane (after
all, if interface/addresses are refreshed that is because they
already exist in the kernel), this patch adds additional methods
to queue requests for interfaces and addresses to providers such
that the responses are, when received, ignored. This is achieved
leveraging the existing infrastructure to skip kernel updates
based on the DPLANE_CTX_FLAG_NO_KERNEL flag.

Note that, this should not affect providers exectuting after the
kernel and that providers executing before may, if desired, reset
that flag.

Signed-off-by: Fredi Raspall <fredi@githedgehog.com>